Separate mock api from webpackconfig
